requirements.txt
flask==2.0.1
python-docx==0.8.11
PyPDF2==3.0.1
beautifulsoup4==4.9.3
requests==2.26.0
openai==1.3.0
python-dotenv==0.19.0
flask-wtf==0.15.1 